- Added a table view controller - Started tweet design
