+++ OpenTweet/TimelineViewController.swift
@@ -9,14 +9,6 @@
 import UIKit
 
 class TimelineViewController: UIViewController {
-    private var userTimeline = UserTimeline(timeline: [])
-
-	override func viewDidLoad() {
-		super.viewDidLoad()
-        if let timeline = TimelineDecoder().decodeTimeline() {
-            userTimeline = timeline
-        }
-        print(userTimeline.timeline)
-	}
+    
 }
